Fix environment
The setup was broken due to env var substitution not working with
empty strings in shell files.

All env vars that are used in shell have default values in the
docker-compose file. Env vars for Django settings can have empty
strings since we take care of that in python and count it as unset.